product bugs fixed

+++ Entities/Concrete/Product.cs
@@ -6,10 +6,10 @@ namespace Entities.Concrete
     {
         public int Id { get; set; }
         public int ModelType { get; set; }
-        public int ItemTypeCode { get; set; }
+        public byte ItemTypeCode { get; set; }
         public string ItemCode { get; set; }
         public string ItemDescription { get; set; }
-        public int ItemDimTypeCode { get; set; }
+        public byte ItemDimTypeCode { get; set; }
         public string UnitOfMeasureCode1 { get; set; }
         public string ItemTaxGrCode { get; set; }
         public int ProductHierarchyID { get; set; }
